backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
opencv-python>=4.8.0
httpx>=0.27.0
Pillow>=10.0.0
numpy>=1.24.0
SpeechRecognition>=3.10.0
openai>=1.0.0
pygame>=2.5.0
firebase-admin>=6.0.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
PyAudio>=0.2.14
PyJWT>=2.8.0
bcrypt>=4.0.0
